Separated NN tests from flow tests
