+++ tests/nn/__init__.py
@@ -1,0 +1,1 @@
+from __future__ import absolute_import, division, print_function
